Refactored account repo to only handle account objects
