+++ src/main/java/com/qa/business/repository/IAccountRepository.java
@@ -1,9 +1,13 @@
 package com.qa.business.repository;
 
+import java.util.List;
+
+import com.qa.persistence.domain.Account;
+
 public interface IAccountRepository {
-	String createAccount(String jsonAccount);
-	String updateAccount(Long id, String jsonAccount);
-	String deleteAccount(Long id);
-	String getAllAccounts();
-	String getAccount(Long id);
+	Account createAccount(Account account);
+	Account updateAccount(Long id, Account updatedAccount);
+	Account deleteAccount(Long id);
+	List<Account> getAllAccounts();
+	Account getAccount(Long id);
 }
